chore: update versions

+++ scripts/generate_versions.ts
@@ -48,11 +48,11 @@ pub enum Versions {
     ${final_list.join("\n    ")}
 }
 
-pub fn version_within_current_patch(version: u16) -> bool {
+pub fn version_within_current_patch(version: u32) -> bool {
     version >= CURRENT_MAJOR && version <= CURRENT_PATCH
 }
 
-pub fn version_within_current_minor(version: u16) -> bool {
+pub fn version_within_current_minor(version: u32) -> bool {
     version >= CURRENT_MAJOR && version <= CURRENT_MINOR
 }`;
 await Deno.writeTextFile("src/mcpe/version.rs", new_file);
